Support separate configuration of multicast logger
This adds support for separating the Logger and MulticastLoggerBackend
configurations so that you can do something like this:
```
config :logger, :logger_multicast_backend,
  target: {{224,0,0,224}, 9999}
```
Configuration by passing all parameters in a tuple to the logger is
still supported:
```
config :logger,
  backends: [ :console,
              {LoggerMulticastBackend,
               target: {{224,0,0,244}, 999},
               format: "$time $metadata[$level] $message\n"} ]
```
This latter configuration takes precedence over the separate
configuration in a similar way to the :console logger.

This change also fixes something that may have been misleading to users.
The `format` option was being set on the main `:logger` configuration
where it has no effect. For it to work, it has to be set in the
multicast backend options.